Add files via upload
The Python file creates a list of strings containing the Latin words contained in loren_Ipsum.txt file, then uses a "Bag of words" approach -- identifying the list of unique words constituting the file and counting number of appearance. It then plots a series of finer and finer histograms, displaying the frequency of appearance of the unique words.